# src/main/resources/openapi/webshop-openapi.yaml
openapi: 3.0.1
info:
  title: Webshop API
  description: Awesome APIs
  version: 1.0.0
paths:

  /products:
    get:
      tags:
        - Products
      summary: Retrieve all products
      operationId: getProducts
      responses:
        200:
          description: Product response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'




  /shoppingcarts:
    get:
      tags:
        - ShoppingCarts
      summary: Retrieve all Shoppingcarts
      operationId: geShoppingCarts
      responses:
        200:
          description: Product response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shoppingcart'
    post:
      tags:
        - ShopingCartCreate
      summary: Create a note
      operationId: createShoppingCart
      requestBody:
        description: Shoppingcart to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Shoppingcart'
        required: true
      responses:
        200:
          description: Shoppingcart response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shoppingcartextens'
  /shoppingcarts/{SoppingcartId}:
    put:
      tags:
        - Product
      summary: add antal product
      operationId: addProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Shoppingcart'
        required: true
      parameters:
        - in: path
          name: SoppingcartId
          schema:
            type: string
            format: uuid
          required: true
          description: ID of the Product
      responses:
        200:
          description: Shopingcart response
          content:
            application/jason:
              schema:
                $ref: '#/components/schemas/Shoppingcartextens'

  /notes:
    get:
      tags:
        - Notes
      summary: Retrieve all notes
      operationId: getNotes
      responses:
        200:
          description: Notes response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NoteEntity'
    post:
      tags:
        - Notes
      summary: Create a note
      operationId: createNote
      requestBody:
        description: Note to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
        required: true
      responses:
        200:
          description: Notes response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteEntity'
  /notes/{noteId}:
    get:
      tags:
        - Notes
      summary: Retrieve a single note
      operationId: getNote
      parameters:
        - in: path
          name: noteId
          schema:
            type: string
            format: uuid
          required: true
          description: ID of the note
      responses:
        200:
          description: Notes response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteEntity'
    put:
      tags:
        - Notes
      summary: Update a note
      operationId: updateNote
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
        required: true
      parameters:
        - in: path
          name: noteId
          schema:
            type: string
            format: uuid
          required: true
          description: ID of the note
      responses:
        200:
          description: Notes response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteEntity'
    delete:
      tags:
        - Notes
      summary: Delete a note
      operationId: deleteNote
      parameters:
        - in: path
          name: noteId
          schema:
            type: string
            format: uuid
          required: true
          description: ID of the note
      responses:
        204:
          description: Deleted
components:
  schemas:
    Note:
      type: object
      description: Representation of a note
        - name
        - description
      properties:
        name:
          type: string
        description:
          type: string
    NoteEntity:
      allOf:
        - $ref: '#/components/schemas/Note'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
              format: uuid
    Product:
      type: object
      description: Product to buy from shop
        - name
        - description
        - price
        - available
        - image
        - id
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        available:
          type: boolean
        image:
          type: string
        id:
          type: string

    Shoppingcart:
      type: object
      properties:
        productItems:
          type: array
          items:
            $ref : '#/components/schemas/Productitems'
    Shoppingcartextens:
      allOf:
        - $ref: '#/components/schemas/Shoppingcart'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
              format: uuid

    Productitems:
      type: object
      properties:
        amount:
          type: integer
        category:
          $ref: '#/components/schemas/Product'


    Sales:
      type: object
      description: Product on sales
        - name
        - description
        - deadline-date
      properties:
        name:
          $ref: '#/components/schemas/Product'
        description:
          type: string
        deadline-date:
          type: string
          format: date